Added test; removed marker file.

+++ recipes/chrome.rb
@@ -1,23 +1,28 @@
-##  This is very experimental and unsupported.  Don't expect it to work,
-##  though it seems to work OK.
+unless File.exists?("/Applications/Google Chrome.app")
 
-run_unless_marker_file_exists("google_chrome") do
-  execute "download chrome to temp dir" do
-    command "curl -o /tmp/googlechrome.dmg #{node["chrome_download_location"]}"
-    user WS_USER
+  remote_file "/tmp/chrome.dmg" do
+    source node["chrome_download_uri"]
+    mode "0644" 
   end
   
   execute "mount chrome dmg" do
-    command "hdid /tmp/googlechrome.dmg"
+    command "hdid /tmp/chrome.dmg"
     user WS_USER
   end
-  
+
   execute "copy chrome to /Applications" do
-    command 'cp -rf /Volumes/Google\\ Chrome/Google\\ Chrome.app /Applications/'
-  end
-  
-  execute "unmount dmg" do
-    command "hdiutil detach  /Volumes/Google\\ Chrome/"
+    command "cp -rf /Volumes/Google\\ Chrome/Google\\ Chrome.app /Applications/"
     user WS_USER
+  end
+
+  execute "unmount dmg" do
+    command "hdiutil detach /Volumes/Google\\ Chrome"
+    user WS_USER
+  end
+
+  ruby_block "test to see if Chrome was installed" do
+    block do
+      raise "Chrome install failed" unless File.exists?("/Applications/Google Chrome.app")
+    end
   end
 end
